New approach to making API calls
* Create any valid object using the same call to the API
* Arrange files to share this code among repos
